+++ lib/logger.rb
@@ -380,7 +380,9 @@ class Logger
   include Severity
 
   # Logging severity threshold (e.g. <tt>Logger::INFO</tt>).
-  attr_reader :level
+  def level
+    Thread.current[@key] || @level
+  end
 
   # Sets the log level; returns +severity+.
   # See {Log Level}[rdoc-ref:Logger@Log+Level].
@@ -395,25 +397,20 @@ class Logger
   # Logger#sev_threshold= is an alias for Logger#level=.
   #
   def level=(severity)
-    if severity.is_a?(Integer)
-      @level = severity
-    else
-      case severity.to_s.downcase
-      when 'debug'
-        @level = DEBUG
-      when 'info'
-        @level = INFO
-      when 'warn'
-        @level = WARN
-      when 'error'
-        @level = ERROR
-      when 'fatal'
-        @level = FATAL
-      when 'unknown'
-        @level = UNKNOWN
-      else
-        raise ArgumentError, "invalid log level: #{severity}"
-      end
+    @level = Severity.coerce(severity)
+  end
+
+  # Adjust the log level during the block execution for the current Thread only
+  #
+  #   logger.with_level(:debug) do
+  #     logger.debug { "Hello" }
+  #   end
+  def with_level(severity, &block)
+    current, Thread.current[@key] = level, Severity.coerce(severity)
+    begin
+      yield
+    ensure
+      Thread.current[@key] = current
     end
   end
 
@@ -583,6 +580,7 @@ class Logger
     self.datetime_format = datetime_format
     self.formatter = formatter
     @logdev = nil
+    @key = "__log_level_#{object_id}"
     if logdev && logdev != File::NULL
       @logdev = LogDevice.new(logdev, shift_age: shift_age,
         shift_size: shift_size,